+++ packages/terra-icon/scripts/src/generate-icons/parseSvg.js
@@ -1,4 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies, compat/compat */
+
 import fs from 'fs';
 import path from 'path';
 import { JSDOM } from 'jsdom';